reworking frontend interfaces with bootstrap.js
I am trying to rework the frontend interfaces to use bootstrap. It appeas to be going okay and the pages are scaling well from mobile to laptop, or more apropriately
from my mobile firefox to mylaptops firefox. I wish I had more devices and browsers to test it on. But if I understand correctly, using the bootstrap framework
will fix most of those compatability issues. Anyways Ihave done 4 and there are 10 more pages to do. I got sidetracked a little on the Index page but that is all good now. Also I edited the logo image in response to G's suggestion. Moved the S slightly to the left so it no longer looks like a snake attacking the bird.