Starting to work
Integrated bootstrap.js into pages and it looks good. Perhaps not as completely as it could
but functional, I think is enough for now. Clearly this needs some vigorous cross browser testing but it works on firfox
which is better than IE but I am happy.
Now I just need to clean up the code and I thank you very much.